backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.2
cryptography==46.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.0
pandas==2.2.0
openpyxl==3.1.5
bcrypt==3.2.0
apscheduler==3.10.4